[SERVICES] Refactor Metadata to Class-Based
The new class-based design provides a more organized and maintainable approach to managing metadata, and it also makes it easier to add newer metadata services/objects in the future. This change will improve the overall structure of the services module and make it more flexible to accommodate future changes.

If the developers still wish to use dictionary-based metadata, consider using `metadata.dict()`.